pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "silentinstallhelper"
version = "0.1.0"
description = "Config-driven helper for silent software installations with progress UI."
readme = "readme.md"
requires-python = ">=3.11"
authors = [
  { name = "Steve" }
]
dependencies = [
  "PySide6>=6.8,<7",
  "PyYAML>=6,<7"
]

[project.optional-dependencies]
dev = [
  "pytest>=8,<9",
  "PyInstaller>=6.10,<7"
]

[project.scripts]
silentinstallhelper = "silent_install_helper.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
